backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
greenlet==3.0.3
supabase==2.4.3
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx-mock==0.27.0
